move fxns out of temp_data_source_functions and delete file
